pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "cryptography>=44.0.0",
    "dependency-injector>=4.45.0",
    "fastapi>=0.115.8",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
